requirements_dev.txt

# Package management
pip
pre-commit

# Development
pytest-dotenv



requests
# Code linting
ruff
setuptools
wheel
